added eyes in floorboard
I think this is going to be so much better - added creepy eyes, this will become the spider that cast a net on the guy. Then when Monty tries to blow up the guy, the spider will snatch the bomb.
So you need to keep the spider inside of the hole before blowing him up.